+++ src/Printed/Bundle/Queue/Command/EnsureVhostExistsCommand.php
@@ -2,19 +2,27 @@
 
 namespace Printed\Bundle\Queue\Command;
 
+use Printed\Bundle\Queue\Service\RabbitMqVhostExistenceEnsurer;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
-use Symfony\Component\DependencyInjection\ContainerAwareTrait;
 use RabbitMq;
 
 /**
  * {@inheritdoc}
  */
-class EnsureVhostExistsCommand extends Command implements ContainerAwareInterface
+class EnsureVhostExistsCommand extends Command
 {
-    use ContainerAwareTrait;
+    /** @var RabbitMqVhostExistenceEnsurer */
+    private $rabbitMqVhostExistenceEnsurer;
+
+    public function __construct(
+        RabbitMqVhostExistenceEnsurer $rabbitMqVhostExistenceEnsurer
+    ) {
+        parent::__construct();
+
+        $this->rabbitMqVhostExistenceEnsurer = $rabbitMqVhostExistenceEnsurer;
+    }
 
     /**
      * {@inheritdoc}
@@ -34,9 +42,7 @@ class EnsureVhostExistsCommand extends Command implements ContainerAwareInterfac
             $output->setVerbosity(OutputInterface::VERBOSITY_VERY_VERBOSE);
         }
 
-        $this->container
-            ->get('printed.bundle.queue.service.rabbit_mq_vhost_existence_ensurer')
-            ->ensure();
+        $this->rabbitMqVhostExistenceEnsurer->ensure();
     }
 
 }