+++ src/Printed/Bundle/Queue/ValueObject/QueueBundleOptions.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Printed\Bundle\Queue\ValueObject;
+
+use Symfony\Component\DependencyInjection\ParameterBag\FrozenParameterBag;
+
+/**
+ * A value object for typehinting only.
+ */
+class QueueBundleOptions extends FrozenParameterBag
+{
+
+}